fix(testing): Rename the bot protection method to `setupClerkTestingToken`
