+++ .changeset/four-teachers-clap.md
@@ -1,5 +1,5 @@
 ---
-'@clerk/testing': minor
+'@clerk/testing': major
 ---
 
 Introduce the `@clerk/testing` package
