+++ packages/testing/src/playwright/index.ts
@@ -1,2 +1,2 @@
 export { clerkSetup } from './setup';
-export { clerkBypassBotProtection } from './bypassBotProtection';
+export { setupClerkTestingToken } from './setupClerkTestingToken';
